Logging is implemented. Every function logs its entry and exit with return value where applicable (unless it's huge, as in the ascii representation of the avatar png file). There's also other logging happening at places that seems like it might be nice to have the information.
When a function fails the return log is logged as WARNING or ERROR. 

All the debug dialog (showInfo calls) have been replace with log
entries. All dialogs to the user are repeated as log entries to help
understant the program flow from the log contents.

There are probably other places that could be logged, but this will
greatly assist with tracking down problems. 

The log entries are dated to the millisecond, so we could also figure
out program timing from this without having to resort to actually
impelenting timing code at the beginning and end of functions/blocks.

I didn't add a menubar item to toggle logging yet, but I will. It also
needs to be saved to the config file.