Added a progress bar to the backlog scoreing utility, score_backlog(), and changed the i limit on the scoring while loop to twice the number of points to score (allowing for 50% of the scoring attempts to fail).
Given that the add-on advertises, on the Anki site, to work with reviews
on any platform, I think there should be an automatic check after a
profile is loaded and synced to see if reviews have been completed on
another device. Then the user should be prompted to sync those reviews
to Habitica. 

The current method requires that the user initiate the database check
manually, which is kind of a pain for those who use desktop Anki less
than alternatives.

I'll implement that next and make it easy to disable with a boolean
variable in case it's undesireable for some reason.